aeerrss

+++ funciones.php
@@ -114,10 +114,6 @@
     function eliminarJuegos($conexion){
         $cantidad = $_POST["cantidad_juegos"];
         $id = $_POST["id_arriendo"];
-        if($cantidad == 0){
-            echo "No hay juegos seleccionados";
-            exit();
-        }
         $j=0;
         $juegos[] = "";
         for($i=0; $i<$cantidad; $i++){
@@ -127,7 +123,10 @@
                 $j++;
             }
         }
-
+        if($j == 0){
+            echo "No hay juegos seleccionados";
+            exit();
+        }
         $estado = 0;
         for($i=0; $i<$j; $i++){
             $sql_eliminar = "DELETE FROM juego_arriendo WHERE juego_arriendo.id_arriendo = '$id' AND juego_arriendo.id_juego = '".$juegos[$i]."' ";
